Remove documentation of passphrase-hashing functions.
crypt.texi itself continues to exist, since it also documents
getentropy and getrandom.

I deleted the paragraph at the beginning of crypt.texi about legal
restrictions on cryptographic software, because after this patchset
the only cryptographic code in glibc itself will be the MD5
implementation used by localedef (see first patch in this series),
which is not exposed to users of the library, and the DES
implementation in sunrpc/, which is also slated for removal (right?)
If this paragraph should be preserved, please let me know.