Get rid of WANT_FLOAT128 usage in floatn.h
This header is installed system-wide. It's not correct to introduce a new
macro WANT_FLOAT128 in this because then we are either forcing the compiler to
make it an inbuilt macro to make glibc expose all float128 functionality, or asking
our clients to -DWANT_FLOAT128 to get float128 functionality in glibc.

Given we are primarily going to have float128 enabled GRTE now, we don't need to have
guards for non-float128 cases.